refactored edit profile routes
