+++ src/db.js
@@ -1,0 +1,38 @@
+const pgp = require("pg-promise")();
+
+const dbConfig = {
+  host: process.env.HOST,
+  port: 5432,
+  database: process.env.POSTGRES_DB,
+  user: process.env.POSTGRES_USER,
+  password: process.env.POSTGRES_PASSWORD,
+};
+
+const db = pgp(dbConfig);
+
+// db test
+db.connect()
+  .then((obj) => {
+    console.log("Database connection successful");
+    obj.done();
+  })
+  .catch((error) => {
+    console.error("Database connection error:", error);
+    console.error("Connection details:", {
+      host: dbConfig.host,
+      port: dbConfig.port,
+      database: dbConfig.database,
+      user: dbConfig.user,
+    });
+  });
+
+// db check if table exists
+db.query("SELECT * FROM job_post")
+  .then((result) => {
+    console.log("Table exists");
+  })
+  .catch((error) => {
+    console.error("Table does not exist:", error);
+  });
+
+module.exports = db;
